ai-skills-extraction/requirements.txt
kafka-python
transformers
redis
torch